requirements.txt
colorama==0.4.6
greenlet==3.1.1
pycryptodome==3.21.0
PyMySQL==1.1.1
PyYAML==6.0.2
SQLAlchemy==2.0.37
typing_extensions==4.12.2
websockets==14.2
